// src/app/models/grid/therapeutic-plan.grid.ts
export interface TherapeuticPlanGrid {
  id: number;
  currentEditorId: string;
  patientId: number;
  patientName: string;
  medicalAppointmentPlanningId: number;
  medicalAppointmentPlanningDescription: string;
  medicalAppointmentId: number;
  medicalAppointmentTeamId: number;
  style: any;
  medicalAppointmentTeamDescription: string;
  medicalAppointmentRoom: string;
  medicalAppointmentConsultDescription: string;
}